requirements.txt
# AutoGen Multi-Agent Code Generation System Dependencies

# Core AutoGen Framework
autogen-agentchat>=0.2.0
autogen-core>=0.2.0
autogen-ext[openai]>=0.2.0

# Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.0

# Data Validation and Configuration
pydantic>=2.0.0
PyYAML>=6.0

# Environment Management
python-dotenv>=1.0.0

# HTTP and Networking
requests>=2.31.0
aiohttp>=3.8.0

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality and Security
ruff>=0.1.0
mypy>=1.5.0
black>=23.0.0

# Development and Debugging
ipython>=8.0.0

# Optional: Additional Model Providers
# Uncomment as needed for specific providers

# For Azure OpenAI
# azure-identity>=1.14.0

# For Google/Vertex AI
# google-cloud-aiplatform>=1.35.0

# For Anthropic Claude
# anthropic>=0.3.0

# For Cohere
# cohere>=4.0.0

# Production Deployment (optional)
# gunicorn>=21.0.0
# gevent>=23.0.0